Update 6:45 PST

+++ src/main/java/dev/canlapan/handlers/EmployeeHandlers/DeleteEmployeesHandler.java
@@ -1,6 +1,7 @@
 package dev.canlapan.handlers.EmployeeHandlers;
 
 import dev.canlapan.app.App;
+import dev.canlapan.entities.Employee;
 import io.javalin.http.Context;
 import io.javalin.http.Handler;
 import org.jetbrains.annotations.NotNull;
@@ -10,15 +11,16 @@ public class DeleteEmployeesHandler implements Handler {
     @Override
     public void handle(@NotNull Context ctx) throws Exception {
         int employeeID = Integer.parseInt(ctx.pathParam("employeeID"));
-        boolean result = App.employeeService.deleteEmployeeID(employeeID);
-        if(result){
-            ctx.status(200);//status code meaning success but nothing is returned
-            ctx.result("Deleted Employee ID " + employeeID);
-            result = false;
-        }else {
+        Employee temp = App.employeeService.retrieveEmployeeByID(employeeID);
+        if (temp == null){
             ctx.status(404);
-            ctx.result("Employee ID " + employeeID +" not found");
-        }
+            ctx.result("Employee ID " + employeeID + " not found");
+            return;
+        }else {
+            App.employeeService.deleteEmployeeID(employeeID);
+            ctx.status(200);
+            ctx.result("Employee has been deleted")
+;        }
     }
 }
